+++ wp-php-console.php
@@ -1,32 +1,30 @@
 <?php
 /**
- * Plugin Name:       WP PHP Console
- * Plugin URI:        https://github.com/unfulvio/wp-php-console/
- * Description:       An implementation of PHP Console for WordPress. Easily debug and trace PHP errors and warnings from your Chrome dev tools console using a Google Chrome extension.
+ * Plugin Name:  WP PHP Console
+ * Plugin URI:   https://github.com/unfulvio/wp-php-console/
+ * Description:  An implementation of PHP Console for WordPress. Easily debug and trace PHP errors and warnings from your Chrome dev tools console using a Google Chrome extension.
  *
- * Version:           1.4.1
+ * Version:      1.5.0
  *
- * Author:            Fulvio Notarstefano
- * Author URI:        https://github.com/unfulvio/
+ * Author:       Fulvio Notarstefano
+ * Author URI:   https://github.com/unfulvio/
  *
- * License:           GPL-2.0+
- * License URI:       http://www.gnu.org/licenses/gpl-2.0.txt
+ * License:      GPL-2.0+
+ * License URI:  http://www.gnu.org/licenses/gpl-2.0.txt
  *
- * Text Domain:       wp-php-console
- * Domain Path:       /languages
+ * Text Domain:  wp-php-console
+ * Domain Path:  /languages
  */
 
-if ( ! defined( 'WPINC' ) ) {
-	exit;
-}
+defined( 'ABSPATH' ) or exit;
 
 /**
  * WP PHP Console
- * Copyright (c) 2014-2015 Fulvio Notarstefano <fulvio.notarstefano@gmail.com>
+ * Copyright (c) 2014-2016 Fulvio Notarstefano <fulvio.notarstefano@gmail.com>
  * and contributors https://github.com/unfulvio/wp-php-console/graphs/contributors
  *
- * PhpConsole server library.
- * Copyright (c) 2011-2013 by Barbushin Sergey <barbushin@gmail.com>
+ * PhpConsole server library
+ * Copyright (c) 2011-2016 by Barbushin Sergey <barbushin@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2 or, at
@@ -44,16 +42,16 @@ if ( ! defined( 'WPINC' ) ) {
  */
 
 // Composer fallback for PHP < 5.3.0.
-if ( version_compare( PHP_VERSION, '5.3.0' ) === -1 ) {
-	include_once 'vendor/autoload_52.php';
+if ( -1 === version_compare( PHP_VERSION, '5.3.0' ) ) {
+	require_once dirname( __FILE__ ) . '/vendor/autoload_52.php';
 } else {
-	include_once 'vendor/autoload.php';
+	require_once dirname( __FILE__ ) . '/vendor/autoload.php';
 }
 
 /**
  * WP PHP Console requires PHP 5.4.0 minimum.
  * @link https://make.wordpress.org/plugins/2015/06/05/policy-on-php-versions/
- * @link https://github.com/nekojira/wp-requirements
+ * @link https://github.com/unfulvio/wp-requirements
  */
 $this_plugin_checks = new WP_Requirements(
 	'WP PHP Console',
@@ -63,12 +61,12 @@ $this_plugin_checks = new WP_Requirements(
 	)
 );
 
-if ( $this_plugin_checks->pass() === false ) {
+if ( false === $this_plugin_checks->pass() ) {
 	// Stop.
 	$this_plugin_checks->halt();
 	return;
 } else {
 	// Load the main class of this plugin.
-	require_once 'includes/class-wp-php-console.php';
-	new \WP_PHP_Console\Plugin();
+	require_once dirname( __FILE__ ) . '/includes/class-wp-php-console.php';
+	return new \WP_PHP_Console\Plugin();
 }